traffic-data/requirements.txt
pandas
duckdb
python-dotenv
requests